move _generate_obj_name to components.py
